Add to_numpy functions for sketch structures.
