+++ python/hll.cpp
@@ -35,6 +35,10 @@ PYBIND11_MODULE(sketch_hll, m) {
             return h != h2;
         }).def("write", [](const sketch::hll_t &h, std::string path) {
             h.write(path);
+        }).def("to_numpy", [](const sketch::hll_t &h) {
+            py::array_t<uint8_t> ret(py::ssize_t(h.size()));
+            std::copy(h.core().data(), h.core().data() + h.size(), (uint8_t *)ret.request().ptr);
+            return ret;
         });
     m.def("jaccard_index", [](hll_t &h1, hll_t &h2) {
             return jaccard_index(h1, h2);
